otl2html: don't abort if the title is followed by an empty line
otl2html beautifies the (possible) title line before using it.  Likewise,
it also beautifies the following line if it is treating the title specially.
However this process will strip the trailing newline character, leading
to an empty string if the line is empty.  Subsequently, it treats empty
strings as end-of-file indications and stops processing.

The fix is not to strip/beautify the line before entering the main
processing loop, since processLine beautifies the lines itself.